Replace singular generate script with Makefile and configure
At first glance this is way more complicated code but it allows us to
easily expanded this with tests.
We need to somehow integrate compilation to tests and tests in general.
In the end it is way easier to use make instead of writing our own shell
script that integrates it all together.