Add tests
These are scripts that simulate updater's script environment. We do not
have to run updater itself thanks to this and we have more control over
possible control paths in scripts.

There is a lot of duplicate code but that is intentional. It is
meant to be in future copied and kept as fixed old implementation. That
decreases possibly of new changes breaking old behavior and decreases
complexity in future (with duplicity as a trade-off).