base-min: remove updater from minimal installation
The minimal installation should be just minimal hardware support. That
does not include updater. It also makes sense to have updater included
in base.lua as that is the base script updater-ng uses.
The full idea of this is that updater can be used to build image and
externally manage root with any base script but only full base should
contain self-updating.