base-min: move opkg from base-min to base
With removed updater from base-min the opkg should be removed as well as
it serves the same purpose and is not essential for hardware
functionality.
Again to recapitulate the idea: The only self-updating and thus self
package installing base is the base. Any other base either have to
include base.lua or root based on it has to be managed externally (out
of root in updater or --dest in opkg).