drivers: add error for invalid usage of drivers lists
This makes the original fatal error that devices passed to pairs is nil
to non-fatal. This also gives much more direct message.

Drivers are important but there are optional drivers and thus having
their invalid usage/setup as fatal makes no sense. They would just not
be included. It is a regular error only.