kmod: add mt7915e as wi-fi driver
In this commit, I add driver mt7915e for
MEDIATEK Corp. MT7915E 802.11ax [14c3:7915] [1], which we tested
and upcoming batch of Turris Omnia routers will include AX Wi-Fi card by default.
This is necessary as file kmod.list is included by base-min and we make
sure that the driver is included in medkits and in the base
installations.

[1] https://www.asiarf.com/shop/hot-sales/wifi6-4t4r-dual-bands-selectable-mpcie-card-ieee802-11ax-ac-a-b-g-n-2-4g-5ghz-aw7915-np1/