repository: add extra argument pkg_hash_required
This increases security by requesting packaages to always have valid
hash in package index. The OPKG does not require this in default so
updater does not as well. At the same time OPKG can change hash and
updater can stay behind and not validating packages at all because of
that. This prevents that situation.

The tests are tweaked to require this argument to be always set.