drivers/usb: move stuff around
The Zigbee and Z-Wave are technically a serial devices but they are
rather separate as they are not intended for connecting serial console
of other device. They use only serial communication. Instead this just
creates them a separate sections.
The serials are also moved after bluetooth because Wi-Fi, Zigbee and
Z-Wave are kind of same so let's have them together.