githooks/pre-push: allow fast-forward of master to develop
The original check was that master can be pushed only with develop
branch. It is what we mostly want but there are cases when we do not
update develop and only master that are valid as well. The example of
such operation is merge of develop to master, which is always
fast-forward due to our workflow.

This halts push only if master commit is not ancestor of develop instead
of halting it always when develop is not being updated.