foris: install foris only if it is required for pakon
This removes explicit request for foris and installs it only if pakon is
requested. Any other request for foris package is now only installed if
foris is going to be.

The stylistic tweak here is removal of condition checks for for_l10n
function. This function is defined in utils.m4 and thus directly in all
files so this condition made no sense.