drivers/pci.lua: fix for Turris RTROM01-WS3294
Unfortunately, this fault is done by the HW report, which I got from our
hardware department. They swapped vendor and device ID in the report and
I could not verify it. Since I did not have the mini PCIe card, I
could check it with the Mediatek MT7612, which we already have here
listed. :-(

Fixes: 01444c9be017b5c7e304a139d7ed97446c499168 ("pci.lua: add support for Turris RTROM01-WS3294")