base-fix: add fix-turris1x-btrfs-sdcard
This applies only to the Turris 1.x and we want to distribute it to all
of those routers. The issue is that there is no package to pick it up
from so we have to use Turris OS version. The currently developed
version is 6.0.0. This makes this fix package kept in system for some
time (till the TOS 6.0.1) but that should be all right.