base-fix: fix cyclic dependency when updating Mox
The Mox and Shields were unable to update due to error:
  Package mox-support is requested as critical. Cyclic dependency is not allowed for critical requests.

The reason for that was this additional dependency that created the
loop. The reasoning for it was to force the removal of
/etc/fw_env.config before we generate the new one by uboot-envtools.
Package that actually contains that file wasn't mox-support but
mox-generic-support and that does not actually depend on uboot-envtools
and thus breaks the loop.