base-min: install Knot Resolver for Turris 1.x routers
We will install both DNS resolvers, but those are not going to start
automatically by themselves, they are managed (started) by resolver-conf.
Even in services wanted (turris-build repository) there are not there,
there is resolver-conf.

We can not remove unbound for Turris 1.x routers right now as there
might be some users, who are using it and if we remove it, they will end
up with broken Internet connectivity.