base-fix: We want to keep the compatibility list longer
Data in the list are interpreted in runtime and not stored anywhere. We
might want to implement a fix package that would rename manually
installed packages, but currently the compatibility list makes things
work good enough, so let's keep it active to let people migrate manually
for the time being.