nas: install transmission's web when transmission is selected
The reason for this change is that currently, Turris WebApps is
installed only when users install transmission-web [1].

Transmission can run without the web ui as we install the LuCI app [2] for
that and some experienced users can use for configuration and managing
transmission CLI.

In OpenWrt packaging, if users install transmission-web, then it pulls
transmission, this is done due to size capacity on routers supported by
OpenWrt and opt-in. [3]

Let's make it more user friendly for our customers and install even
transmission-web, so they have it with WebApps integration as well.

[1] https://gitlab.nic.cz/turris/os/updater-lists/-/blob/248c5cb765df6cead5bbf88bcd523fa074d3a9ed/webapps.lua.m4#L9
[2] https://github.com/openwrt/luci/blob/5455a54d1cdb08065eca0afe3a47d1ba84c46049/applications/luci-app-transmission/Makefile#L10
[3] https://github.com/openwrt/packages/blob/5dffc0af0e8f7b8e609c309985ea23e95ed46fef/net/transmission/Makefile#L63